refactor: apply mission 2
