+++ src/main/java/kitchenpos/eatinorders/domain/Order.java
@@ -12,6 +12,8 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
+import kitchenpos.eatinorders.domain.constant.OrderStatus;
+import kitchenpos.eatinorders.domain.constant.OrderType;
 
 import java.time.LocalDateTime;
 import java.util.List;
