+++ src/main/java/kitchenpos/eatinorders/domain/constant/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.domain.constant;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
