+++ src/main/java/kitchenpos/menus/repository/JpaMenuRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.repository;
 
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
