+++ src/main/java/kitchenpos/products/repository/JpaProductRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.repository;
 
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
