+++ src/test/java/kitchenpos/eatinorders/application/InMemoryOrderRepository.java
@@ -2,7 +2,7 @@ package kitchenpos.eatinorders.application;
 
 import kitchenpos.eatinorders.domain.Order;
 import kitchenpos.eatinorders.domain.OrderRepository;
-import kitchenpos.eatinorders.domain.OrderStatus;
+import kitchenpos.eatinorders.domain.constant.OrderStatus;
 import kitchenpos.eatinorders.domain.OrderTable;
 
 import java.util.ArrayList;
